// src/constants/order-status-enum.ts
export enum OrderStatusEnum {
   REQUESTED = 'REQUESTED',
   DELIVERY_IN_TRANSIT = 'DELIVERY_IN_TRANSIT',
   COLLECTED_FOR_DELIVERY = 'COLLECTED_FOR_DELIVERY',
   ORDER_DELIVERED = 'ORDER_DELIVERED',
   COLLECTION_CENTER = 'COLLECTION_CENTER',
   COLLECTED = 'COLLECTED',
   DELIVERY_FAILED = 'DELIVERY_FAILED',
}